have a test only for the plugin
