+++ test/e2e/v3/generate_test.go
@@ -34,7 +34,6 @@ import (
 	"sigs.k8s.io/kubebuilder/v3/test/e2e/utils"
 )
 
-//nolint:dupl
 // GenerateV2 implements a go/v2 plugin project defined by a TestContext.
 func GenerateV2(kbc *utils.TestContext) {
 	var err error
@@ -130,7 +129,6 @@ Count int `+"`"+`json:"count,omitempty"`+"`"+`
 #    name: webhook-service`, "#")).To(Succeed())
 }
 
-//nolint:dupl
 // GenerateV3 implements a go/v3(-alpha) plugin project defined by a TestContext.
 func GenerateV3(kbc *utils.TestContext, crdAndWebhookVersion string, restrictive bool) {
 	var err error
@@ -272,101 +270,6 @@ func uncommentPodStandards(kbc *utils.TestContext) {
           type: RuntimeDefault`); err == nil {
 		ExpectWithOffset(1, err).NotTo(HaveOccurred())
 	}
-}
-
-//nolint:dupl
-// GenerateV3WithDeployImage implements a go/v3 plugin and the deployImage one
-func GenerateV3WithDeployImage(kbc *utils.TestContext) {
-	var err error
-
-	By("initializing a project with go/v3")
-	err = kbc.Init(
-		"--plugins", "go/v3",
-		"--project-version", "3",
-		"--domain", kbc.Domain,
-		"--fetch-deps=false",
-	)
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-
-	By("creating API definition with deploy-image/v1-alpha plugin")
-	err = kbc.CreateAPI(
-		"--group", kbc.Group,
-		"--version", kbc.Version,
-		"--kind", kbc.Kind,
-		"--plugins", "deploy-image/v1-alpha",
-		"--image", "memcached:1.6.15-alpine",
-	)
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-
-	By("implementing the API without the plugin")
-	ExpectWithOffset(1, pluginutil.InsertCode(
-		filepath.Join(kbc.Dir, "api", kbc.Version, fmt.Sprintf("%s_types.go", strings.ToLower(kbc.Kind))),
-		fmt.Sprintf(`type %sSpec struct {
-`, kbc.Kind),
-		`	// +optional
-Count int `+"`"+`json:"count,omitempty"`+"`"+`
-`)).Should(Succeed())
-
-	By("scaffolding mutating and validating webhooks")
-	err = kbc.CreateWebhook(
-		"--group", kbc.Group,
-		"--version", kbc.Version,
-		"--kind", kbc.Kind,
-		"--defaulting",
-		"--programmatic-validation",
-	)
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-
-	By("implementing the mutating and validating webhooks")
-	err = pluginutil.ImplementWebhooks(filepath.Join(
-		kbc.Dir, "api", kbc.Version,
-		fmt.Sprintf("%s_webhook.go", strings.ToLower(kbc.Kind))))
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-
-	By("uncomment kustomization.yaml to enable webhook and ca injection")
-	ExpectWithOffset(1, pluginutil.UncommentCode(
-		filepath.Join(kbc.Dir, "config", "default", "kustomization.yaml"),
-		"#- ../webhook", "#")).To(Succeed())
-	ExpectWithOffset(1, pluginutil.UncommentCode(
-		filepath.Join(kbc.Dir, "config", "default", "kustomization.yaml"),
-		"#- ../certmanager", "#")).To(Succeed())
-	ExpectWithOffset(1, pluginutil.UncommentCode(
-		filepath.Join(kbc.Dir, "config", "default", "kustomization.yaml"),
-		"#- ../prometheus", "#")).To(Succeed())
-	ExpectWithOffset(1, pluginutil.UncommentCode(
-		filepath.Join(kbc.Dir, "config", "default", "kustomization.yaml"),
-		"#- manager_webhook_patch.yaml", "#")).To(Succeed())
-	ExpectWithOffset(1, pluginutil.UncommentCode(
-		filepath.Join(kbc.Dir, "config", "default", "kustomization.yaml"),
-		"#- webhookcainjection_patch.yaml", "#")).To(Succeed())
-	ExpectWithOffset(1, pluginutil.UncommentCode(filepath.Join(kbc.Dir, "config", "default", "kustomization.yaml"),
-		`#- name: CERTIFICATE_NAMESPACE # namespace of the certificate CR
-#  objref:
-#    kind: Certificate
-#    group: cert-manager.io
-#    version: v1
-#    name: serving-cert # this name should match the one in certificate.yaml
-#  fieldref:
-#    fieldpath: metadata.namespace
-#- name: CERTIFICATE_NAME
-#  objref:
-#    kind: Certificate
-#    group: cert-manager.io
-#    version: v1
-#    name: serving-cert # this name should match the one in certificate.yaml
-#- name: SERVICE_NAMESPACE # namespace of the service
-#  objref:
-#    kind: Service
-#    version: v1
-#    name: webhook-service
-#  fieldref:
-#    fieldpath: metadata.namespace
-#- name: SERVICE_NAME
-#  objref:
-#    kind: Service
-#    version: v1
-#    name: webhook-service`, "#")).To(Succeed())
-
 }
 
 // GenerateV3 implements a go/v3(-alpha) plugin project defined by a TestContext.
